+++ packages/hero_ui/lib/src/components/components.dart
@@ -3,6 +3,7 @@
 export 'button/button.dart';
 export 'button_group/button_group.dart';
 export 'close_button/close_button.dart';
+export 'kbd/kbd.dart';
 export 'separator/separator.dart';
 export 'spinner/spinner.dart';
 export 'typography/typography.dart';